add category servlets

+++ src/main/java/com/example/posts/Dao/PostDao.java
@@ -2,6 +2,6 @@ package com.example.posts.Dao;
 
 import com.example.posts.model.Post;
 
-public interface PostDao extends GenericDao<Post,Long> {
+public interface PostDao extends GenericDao<Post,Integer> {
 
 }
